refactor(PlayersPanel): organizes PlayersPanel and subcomponents files, adds new shine effect for when user input is required, enhanced text, butons and total points styles
